% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_plt.R
\name{combine_plt}
\alias{combine_plt}
\title{Combine plot files}
\usage{
combine_plt(files, out = "int.dat", type = "short")
}
\arguments{
\item{files}{Path plot files}

\item{out}{File output}

\item{type}{Type plot file}
}
\value{
A vector of characters
}
\description{
Combine multiple plot files in a single file
}
\examples{
file_dat <- system.file("extdata", "rank.dat", package="Rcalmet")
file_out <- tempfile()
plt <- combine_plt(c(file_dat, file_dat), type="long")
# writeLines(plt, file_out)

}
