% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTSF.R
\name{importTSF}
\alias{importTSF}
\title{Import data of TSF file, output METSERIES}
\usage{
importTSF(tsf, type)
}
\arguments{
\item{tsf}{Path TSF file}

\item{type}{Type output TSF file. "m3d" or "calmet"}
}
\value{
A data.frame
}
\description{
Import data to TSF file generaded for METSERIES as a data.frame.
}
\examples{
file_tsf <- system.file("extdata", "pasco.tsf", package="calmet")
importTSF(file_tsf, "m3d")

}
