% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importUA.R
\name{importUA}
\alias{importUA}
\title{Import data of UA file, output UAMAKE}
\usage{
importUA(ua, lvs)
}
\arguments{
\item{ua}{Path UAMAKE output file}

\item{lvs}{Levels}
}
\value{
A list
}
\description{
Combine multiple plot files
}
\examples{
file_dat <- system.file("extdata", "uamake.dat", package="Rcalmet")
importUA(file_dat, 22)

}
